backend/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
